requirements.txt
Flask
psycopg2-binary
qrcode
gunicorn
python-dotenv
Pillow